refactor the commands setup
