+++ gquantlab/src/index.ts
@@ -4,13 +4,13 @@ import {
   ILayoutRestorer
 } from '@jupyterlab/application';
 
-import { IFileBrowserFactory, FileDialog } from '@jupyterlab/filebrowser';
+import { IFileBrowserFactory } from '@jupyterlab/filebrowser';
 
 import { ILauncher } from '@jupyterlab/launcher';
 
 import { IMainMenu } from '@jupyterlab/mainmenu';
 
-import { Token } from '@lumino/coreutils';
+import { Token, ReadonlyJSONObject } from '@lumino/coreutils';
 
 import { MODULE_NAME, MODULE_VERSION } from './version';
 import * as widgetExports from './widget';
@@ -30,8 +30,6 @@ import {
   notebookIcon
 } from '@jupyterlab/ui-components';
 
-import { folderIcon } from '@jupyterlab/ui-components';
-
 import { ToolbarButton } from '@jupyterlab/apputils';
 
 import { toArray } from '@lumino/algorithm';
@@ -40,29 +38,36 @@ import {
   IWidgetTracker,
   WidgetTracker
 } from '@jupyterlab/apputils';
-import {
-  GquantWidget,
-  GquantFactory,
-  IAllNodes,
-  INode,
-  IChartInput
-} from './document';
+import { GquantWidget, GquantFactory, IAllNodes } from './document';
 import { Menu, Widget } from '@lumino/widgets';
 import { IJupyterWidgetRegistry } from '@jupyter-widgets/base';
 import { INotebookTracker } from '@jupyterlab/notebook';
 import { CodeCell } from '@jupyterlab/cells';
-import { MainView, OUTPUT_COLLECTOR, OUTPUT_TYPE } from './mainComponent';
-import YAML from 'yaml';
-import { EditorPanel } from './EditorPanel';
+import { MainView } from './mainComponent';
 import '../style/editor.css';
 import { DocumentRegistry } from '@jupyterlab/docregistry';
+import {
+  COMMAND_CONVERT_CELL_TO_FILE,
+  setupCommands,
+  COMMAND_OPEN_NEW_FILE,
+  COMMAND_NEW_TASK_GRAPH,
+  COMMAND_NEW_OPEN_TASK_GRAPH,
+  COMMAND_OPEN_EDITOR,
+  COMMAND_OPEN_NEW_NOTEBOOK,
+  COMMAND_ADD_NODE,
+  COMMAND_RELAYOUT,
+  COMMAND_CHANGE_ASPECT_RATIO,
+  COMMAND_EXECUTE,
+  COMMAND_CLEAN,
+  COMMAND_ADD_OUTPUT_COLLECTOR
+} from './commands';
 
 // import { LabIcon } from '@jupyterlab/ui-components';
 // import { LabIcon } from '@jupyterlab/ui-components/lib/icon/labicon';
 
 //const WIDGET_VIEW_MIMETYPE = 'application/gquant-taskgraph';
 
-const FACTORY = 'GQUANTLAB';
+export const FACTORY = 'GQUANTLAB';
 
 type IGQUANTTracker = IWidgetTracker<GquantWidget>;
 
@@ -125,7 +130,7 @@ function outputEnabled(outputArea: any): boolean {
   return true;
 }
 
-function isLinkedView(currentWidget: any): boolean {
+export function isLinkedView(currentWidget: any): boolean {
   if (currentWidget) {
     for (const view of toArray(currentWidget.children())) {
       if (view instanceof Widget) {
@@ -144,7 +149,7 @@ function isLinkedView(currentWidget: any): boolean {
   return false;
 }
 
-function isEnabled(cell: any): boolean {
+export function isEnabled(cell: any): boolean {
   if (!cell) {
     return false;
   }
@@ -215,11 +220,11 @@ function activateFun(
       };
 
       const notebookCallback = (): void => {
-        app.commands.execute('gquant:openAnNewNotebook');
+        app.commands.execute(COMMAND_OPEN_NEW_NOTEBOOK);
       };
 
       const editorCallback = (): void => {
-        app.commands.execute('gquant:openeditor');
+        app.commands.execute(COMMAND_OPEN_EDITOR);
       };
 
       const layout = new ToolbarButton({
@@ -331,13 +336,6 @@ function activateFun(
     return gquantVisible;
   }
 
-  /**
-   * Whether there is an active graph editor
-   */
-  function isVisible(): boolean {
-    return isGquantVisible() || isCellVisible();
-  }
-
   // Handle state restoration.
   restorer.restore(tracker, {
     command: 'docmanager:open',
@@ -366,390 +364,18 @@ function activateFun(
     fileFormat: 'text'
   });
 
-  // Function to create a new untitled diagram file, given
-  // the current working directory.
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const createGQFile = async (cwd: string) => {
-    const model = await commands.execute('docmanager:new-untitled', {
-      path: cwd,
-      type: 'file',
-      ext: '.gq.yaml'
-    });
-    return commands.execute('docmanager:open', {
-      path: model.path,
-      factory: FACTORY
-    });
-  };
-
-  const convertToGQFile = async (cwd: string): Promise<void> => {
-    const model = await commands.execute('docmanager:new-untitled', {
-      path: cwd,
-      type: 'file',
-      ext: '.gq.yaml'
-    });
-    const mainView = getMainView();
-    const obj = mainView.contentHandler.privateCopy.get('value');
-    model.content = YAML.stringify(obj);
-    model.format = 'text';
-    app.serviceManager.contents.save(model.path, model);
-  };
-
-  commands.addCommand('gquant:convertCellToFile', {
-    label: 'Create Taskgraph from this Cell',
-    caption: 'Create Taskgraph from this Cell',
-    icon: gqIcon,
-    execute: () => {
-      //const cwd = notebookTracker.currentWidget.context.path;
-      const cwd = browserFactory.defaultBrowser.model.path;
-      return convertToGQFile(cwd);
-    },
-    isVisible: isCellVisible
-  });
-
-  commands.addCommand('gquant:toolbarConvertCellToFile', {
-    label: '',
-    caption: 'Create Taskgraph from this Cell',
-    icon: gqIcon,
-    execute: async (args: any) => {
-      //const cwd = notebookTracker.currentWidget.context.path;
-      const cwd = browserFactory.defaultBrowser.model.path;
-      return convertToGQFile(cwd);
-    }
-  });
-
-  commands.addCommand('gquant:selectTheFile', {
-    label: 'Select the file',
-    caption: 'Select the file',
-    icon: folderIcon,
-    execute: async (args: any) => {
-      let dialog = null;
-      if (args && 'filter' in args) {
-        dialog = FileDialog.getOpenFiles({
-          manager: browserFactory.defaultBrowser.model.manager, // IDocumentManager
-          title: 'Select the File',
-          filter: model =>
-            args['filter'].some((d: string): boolean => model.path.endsWith(d))
-        });
-      } else {
-        dialog = FileDialog.getOpenFiles({
-          manager: browserFactory.defaultBrowser.model.manager, // IDocumentManager
-          title: 'Select the File'
-        });
-      }
-      const result = await dialog;
-      if (result.button.accept) {
-        // console.log(result.value);
-        const values = result.value;
-        if (values.length === 1) {
-          return values[0];
-        }
-        //let files = result.value;
-      }
-    }
-  });
-
-  commands.addCommand('gquant:selectThePath', {
-    label: 'Select the Path',
-    caption: 'Select the Path',
-    icon: folderIcon,
-    execute: async () => {
-      const dialog = FileDialog.getExistingDirectory({
-        manager: browserFactory.defaultBrowser.model.manager, // IDocumentManager
-        title: 'Select the Path'
-      });
-      const result = await dialog;
-      if (result.button.accept) {
-        // console.log(result.value);
-        const values = result.value;
-        if (values.length === 1) {
-          return values[0];
-        }
-        //let files = result.value;
-      }
-    }
-  });
-
-  commands.addCommand('gquant:openNewFile', {
-    label: 'Open TaskGraph file',
-    caption: 'Open TaskGraph file',
-    icon: folderIcon,
-    execute: async () => {
-      const dialog = FileDialog.getOpenFiles({
-        manager: browserFactory.defaultBrowser.model.manager, // IDocumentManager
-        filter: model => model.path.endsWith('.gq.yaml')
-      });
-      const result = await dialog;
-      if (result.button.accept) {
-        // console.log(result.value);
-        const values = result.value;
-        if (values.length === 1) {
-          // only 1 file is allowed
-          const payload = { path: values[0].path };
-          const workflows: IChartInput = await requestAPI<any>(
-            'load_graph_path',
-            {
-              body: JSON.stringify(payload),
-              method: 'POST'
-            }
-          );
-          const mainView = getMainView();
-          mainView.contentHandler.contentReset.emit(workflows);
-        }
-        //let files = result.value;
-      }
-    },
-    isVisible: isCellVisible
-  });
-
-  commands.addCommand('gquant:toolbaropenTaskGraph', {
-    label: '',
-    caption: 'Open TaskGraph file',
-    icon: folderIcon,
-    mnemonic: 0,
-    execute: async () => {
-      commands.execute('gquant:openNewFile');
-    }
-  });
-
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const createNewTaskgraph = async (cwd: string) => {
-    const model = await commands.execute('docmanager:new-untitled', {
-      path: cwd,
-      type: 'file',
-      ext: '.gq.yaml'
-    });
-    //let wdg = app.shell.currentWidget as any;
-    // wdg.getSVG()
-    model.content = '';
-    model.format = 'text';
-    app.serviceManager.contents.save(model.path, model);
-  };
-
-  function uuidv4(): string {
-    return Math.random()
-      .toString(36)
-      .substring(2, 15);
-  }
-
-  app.commands.addCommand('gquant:openeditor', {
-    label: 'Task Node Editor',
-    caption: 'Open the Task Node Editor',
-    icon: editIcon,
-    mnemonic: 0,
-    execute: () => {
-      if (isCellVisible()) {
-        if (isEnabled(notebookTracker.activeCell)) {
-          const mainView = getMainView();
-          let panel = null;
-          for (const view of toArray<Widget>(app.shell.widgets())) {
-            if (
-              view instanceof EditorPanel &&
-              (view as EditorPanel).handler === mainView.contentHandler
-            ) {
-              // console.log('found it');
-              panel = view;
-              break;
-            }
-          }
-          if (panel === null) {
-            panel = new EditorPanel(mainView.contentHandler);
-            panel.id = panel.id + uuidv4();
-            app.shell.add(panel, 'main', { mode: 'split-bottom' });
-          } else {
-            app.shell.activateById(panel.id);
-          }
-        } else if (isLinkedView(app.shell.currentWidget)) {
-          const mainView = getMainView();
-          let panel = null;
-          for (const view of toArray<Widget>(app.shell.widgets())) {
-            if (
-              view instanceof EditorPanel &&
-              (view as EditorPanel).handler === mainView.contentHandler
-            ) {
-              // console.log('found it');
-              panel = view;
-              break;
-            }
-          }
-          if (panel === null) {
-            panel = new EditorPanel(mainView.contentHandler);
-            panel.id = panel.id + uuidv4();
-            app.shell.add(panel, 'main', { mode: 'split-bottom' });
-          } else {
-            app.shell.activateById(panel.id);
-          }
-        }
-      } else {
-        const wdg = app.shell.currentWidget as any;
-        wdg.contentHandler.reLayoutSignal.emit();
-        let panel = null;
-        for (const view of toArray<Widget>(app.shell.widgets())) {
-          if (
-            view instanceof EditorPanel &&
-            (view as EditorPanel).handler === wdg.contentHandler
-          ) {
-            // console.log('found it');
-            panel = view;
-            break;
-          }
-        }
-        if (panel === null) {
-          panel = new EditorPanel(wdg.contentHandler);
-          panel.id = panel.id + uuidv4();
-          app.shell.add(panel, 'main', { mode: 'split-bottom' });
-        } else {
-          app.shell.activateById(panel.id);
-        }
-      }
-    },
-    isVisible
-  });
-
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const createNewNotebook = async (input1: string, input2: string) => {
-    const model = await commands.execute('docmanager:new-untitled', {
-      type: 'notebook'
-    });
-    const empty: any[] = [];
-    const notebook = {
-      cells: [
-        {
-          cell_type: "code",
-          execution_count: 1,
-          metadata: {},
-          outputs: empty,
-          source: input1
-        },
-        {
-          cell_type: "code",
-          execution_count: 2,
-          metadata: {},
-          outputs: empty,
-          source: input2
-        }
-      ],
-      metadata: {
-        kernelspec: {
-          display_name: 'Python 3',
-          language: 'python',
-          name: 'python3'
-        }
-      },
-      nbformat: 4,
-      nbformat_minor: 4
-    };
-
-    model.content = notebook;
-    model.format = 'text';
-    const savedModel = await app.serviceManager.contents.save(
-      model.path,
-      model
-    );
-    browserFactory.defaultBrowser.model.manager.open(savedModel.path);
-  };
-
-  // Add a command for creating a new diagram file.
-  commands.addCommand('gquant:create-new', {
-    label: 'TaskGraph',
-    icon: gqIcon,
-    caption: 'Create a new task graph file',
-    execute: () => {
-      const cwd = browserFactory.defaultBrowser.model.path;
-      return createGQFile(cwd);
-    }
-  });
-
-  commands.addCommand('gquant:export-yaml', {
-    label: 'Create an empty Taskgraph',
-    caption: 'Create an empty Taskgraph',
-    execute: () => {
-      const cwd = browserFactory.defaultBrowser.model.path;
-      return createNewTaskgraph(cwd);
-    }
-  });
-
-  commands.addCommand('gquant:reLayout', {
-    label: 'Taskgraph Nodes Auto Layout',
-    caption: 'Taskgraph Nodes Auto Layout',
-    icon: layoutIcon,
-    mnemonic: 0,
-    execute: () => {
-      if (isCellVisible()) {
-        if (isEnabled(notebookTracker.activeCell)) {
-          const mainView = getMainView();
-          mainView.contentHandler.reLayoutSignal.emit();
-        } else if (isLinkedView(app.shell.currentWidget)) {
-          const mainView = getMainView();
-          mainView.contentHandler.reLayoutSignal.emit();
-        }
-      } else {
-        const wdg = app.shell.currentWidget as any;
-        wdg.contentHandler.reLayoutSignal.emit();
-      }
-    },
-    isVisible
-  });
-
-  commands.addCommand('gquant:toolbarReLayout', {
-    label: '',
-    caption: 'Taskgraph Nodes Auto Layout',
-    icon: layoutIcon,
-    mnemonic: 0,
-    execute: () => {
-      commands.execute('gquant:reLayout');
-    }
-  });
-
-  commands.addCommand('gquant:aspect1', {
-    label: 'AspectRatio 1.0',
-    caption: 'AspectRatio 1.0',
-    mnemonic: 14,
-    execute: () => {
-      const mainView = getMainView();
-      mainView.contentHandler.aspectRatio = 1.0;
-      mainView.mimerenderWidgetUpdateSize();
-    },
-    isEnabled: isCellVisible
-  });
-
-  commands.addCommand('gquant:aspect0.3', {
-    label: 'AspectRatio 0.3',
-    caption: 'AspectRatio 0.3',
-    mnemonic: 14,
-    execute: () => {
-      const mainView = getMainView();
-      mainView.contentHandler.aspectRatio = 0.3;
-      mainView.mimerenderWidgetUpdateSize();
-    },
-    isEnabled: isCellVisible
-  });
-
-  commands.addCommand('gquant:aspect0.5', {
-    label: 'AspectRatio 0.5',
-    caption: 'AspectRatio 0.5',
-    mnemonic: 14,
-    execute: () => {
-      const mainView = getMainView();
-      mainView.contentHandler.aspectRatio = 0.5;
-      mainView.mimerenderWidgetUpdateSize();
-    },
-    isEnabled: isCellVisible
-  });
-
-  commands.addCommand('gquant:aspect0.7', {
-    label: 'AspectRatio 0.7',
-    caption: 'AspectRatio 0.7',
-    mnemonic: 14,
-    execute: () => {
-      const mainView = getMainView();
-      mainView.contentHandler.aspectRatio = 0.7;
-      mainView.mimerenderWidgetUpdateSize();
-    },
-    isEnabled: isCellVisible
-  });
+  setupCommands(
+    commands,
+    app,
+    getMainView,
+    browserFactory,
+    isCellVisible,
+    isGquantVisible,
+    notebookTracker
+  );
 
   app.contextMenu.addItem({
-    command: 'gquant:openeditor',
+    command: COMMAND_OPEN_EDITOR,
     selector: '.jp-GQuant'
   });
 
@@ -764,17 +390,17 @@ function activateFun(
   });
 
   app.contextMenu.addItem({
-    command: 'gquant:convertCellToFile',
+    command: COMMAND_CONVERT_CELL_TO_FILE,
     selector: '.jp-GQuant'
   });
 
   app.contextMenu.addItem({
-    command: 'gquant:openNewFile',
+    command: COMMAND_OPEN_NEW_FILE,
     selector: '.jp-GQuant'
   });
 
   app.contextMenu.addItem({
-    command: 'gquant:reLayout',
+    command: COMMAND_RELAYOUT,
     selector: '.jp-GQuant'
   });
 
@@ -794,26 +420,17 @@ function activateFun(
       submenu.title.mnemonic = 0;
       for (let i = 0; i < allNodes[k].length; i++) {
         const name = allNodes[k][i].type;
-        const commandName = `addnode:${k}.${name}`;
-        commands.addCommand(commandName, {
-          label: 'Add ' + name,
-          mnemonic: 4,
-          execute: () => {
-            if (isCellVisible()) {
-              const mainView = getMainView();
-              mainView.contentHandler.nodeAddedSignal.emit(allNodes[k][i]);
-            } else {
-              const wdg = app.shell.currentWidget as any;
-              wdg.contentHandler.nodeAddedSignal.emit(allNodes[k][i]);
-            }
-          },
-          isVisible
+        const args = {
+          name: name,
+          node: (allNodes[k][i] as unknown) as ReadonlyJSONObject
+        };
+        submenu.addItem({
+          command: COMMAND_ADD_NODE,
+          args: args
         });
-        submenu.addItem({ command: commandName });
         if (palette) {
-          const args = { format: 'YAML', label: 'YAML', isPalette: true };
           palette.addItem({
-            command: commandName,
+            command: COMMAND_ADD_NODE,
             category: 'Add New Nodes',
             args: args
           });
@@ -837,109 +454,18 @@ function activateFun(
     selector: '.jp-GQuant'
   });
 
-  const commandExecute = 'gquant:execute';
-  const commandClean = 'gquant:cleanResult';
-
-  commands.addCommand(commandExecute, {
-    label: 'Run',
-    caption: 'Run',
-    icon: runIcon,
-    mnemonic: 0,
-    execute: async () => {
-      if (isCellVisible()) {
-        if (isEnabled(notebookTracker.activeCell)) {
-          const mainView = getMainView();
-          mainView.contentHandler.saveCache.emit();
-          mainView.contentHandler.runGraph.emit();
-        } else if (isLinkedView(app.shell.currentWidget)) {
-          const mainView = getMainView();
-          mainView.contentHandler.saveCache.emit();
-          mainView.contentHandler.runGraph.emit();
-        }
-      }
-    },
-    isVisible: isCellVisible
-  });
-
-  commands.addCommand('gquant:toolbarexecute', {
-    label: '',
-    caption: 'Run',
-    icon: runIcon,
-    mnemonic: 0,
-    execute: async () => {
-      commands.execute(commandExecute);
-    }
-  });
-
-  commands.addCommand(commandClean, {
-    label: 'Clean Result',
-    caption: 'Clean Result',
-    icon: cleanIcon,
-    mnemonic: 0,
-    execute: async () => {
-      if (isCellVisible()) {
-        if (isEnabled(notebookTracker.activeCell)) {
-          const mainView = getMainView();
-          mainView.contentHandler.cleanResult.emit();
-        }
-      }
-    },
-    isVisible: isCellVisible
-  });
-
-  commands.addCommand('gquant:toolbarcleanResult', {
-    label: '',
-    caption: 'Clean Result',
-    icon: cleanIcon,
-    mnemonic: 0,
-    execute: async () => {
-      commands.execute(commandClean);
-    }
-  });
-
   app.contextMenu.addItem({
-    command: commandExecute,
+    command: COMMAND_EXECUTE,
     selector: '.jp-GQuant'
   });
 
   app.contextMenu.addItem({
-    command: commandClean,
+    command: COMMAND_CLEAN,
     selector: '.jp-GQuant'
   });
 
-  commands.addCommand('add:outputCollector', {
-    label: 'Add Output Collector',
-    mnemonic: 1,
-    execute: () => {
-      const output: INode = {
-        id: OUTPUT_COLLECTOR,
-        width: 160,
-        type: OUTPUT_TYPE,
-        conf: {},
-        required: {},
-        // eslint-disable-next-line @typescript-eslint/camelcase
-        output_columns: [],
-        outputs: [],
-        schema: {},
-        ui: {},
-        inputs: [{ name: 'in1', type: ['any'] }]
-      };
-      if (isCellVisible()) {
-        if (isEnabled(notebookTracker.activeCell)) {
-          const mainView = getMainView();
-          mainView.contentHandler.nodeAddedSignal.emit(output);
-        }
-      }
-      if (isGquantVisible()) {
-        const wdg = app.shell.currentWidget as any;
-        wdg.contentHandler.nodeAddedSignal.emit(output);
-      }
-    },
-    isVisible: isVisible
-  });
-
   app.contextMenu.addItem({
-    command: 'add:outputCollector',
+    command: COMMAND_ADD_OUTPUT_COLLECTOR,
     selector: '.jp-GQuant'
   });
 
@@ -951,10 +477,14 @@ function activateFun(
   const submenu = new Menu({ commands });
   submenu.title.label = 'Change Aspect Ratio';
   submenu.title.mnemonic = 0;
-  submenu.addItem({ command: 'gquant:aspect0.3' });
-  submenu.addItem({ command: 'gquant:aspect0.5' });
-  submenu.addItem({ command: 'gquant:aspect0.7' });
-  submenu.addItem({ command: 'gquant:aspect1' });
+
+  [0.3, 0.5, 0.7, 1.0].forEach(d => {
+    submenu.addItem({
+      command: COMMAND_CHANGE_ASPECT_RATIO,
+      args: { aspect: d }
+    });
+  });
+
   app.contextMenu.addItem({
     type: 'submenu',
     submenu: submenu,
@@ -964,7 +494,7 @@ function activateFun(
   // Add a launcher item if the launcher is available.
   if (launcher) {
     launcher.add({
-      command: 'gquant:create-new',
+      command: COMMAND_NEW_OPEN_TASK_GRAPH,
       rank: 1,
       category: 'Other'
     });
@@ -972,151 +502,112 @@ function activateFun(
 
   if (menu) {
     // Add new text file creation to the file menu.
-    menu.fileMenu.newMenu.addGroup([{ command: 'gquant:create-new' }], 40);
+    menu.fileMenu.newMenu.addGroup(
+      [{ command: COMMAND_NEW_OPEN_TASK_GRAPH }],
+      40
+    );
     //palette.addItem({ command: 'gquant:export-yaml', category: 'Notebook Operations', args: args });
-    menu.fileMenu.addGroup([{ command: 'gquant:export-yaml' }], 40);
+    menu.fileMenu.addGroup([{ command: COMMAND_NEW_TASK_GRAPH }], 40);
     menu.addMenu(addNodeMenu, { rank: 40 });
   }
 
   if (palette) {
     const args = { format: 'YAML', label: 'YAML', isPalette: true };
     palette.addItem({
-      command: 'gquant:export-yaml',
+      command: COMMAND_NEW_TASK_GRAPH,
       category: 'Notebook Operations',
       args: args
     });
     palette.addItem({
-      command: 'gquant:reLayout',
+      command: COMMAND_RELAYOUT,
       category: 'GquantLab',
       args: args
     });
-    palette.addItem({
-      command: 'gquant:aspect0.3',
-      category: 'GquantLab',
-      args: args
+    [0.3, 0.5, 0.7, 1.0].forEach(d => {
+      palette.addItem({
+        command: COMMAND_CHANGE_ASPECT_RATIO,
+        category: 'GquantLab',
+        args: { aspect: d }
+      });
     });
     palette.addItem({
-      command: 'gquant:aspect0.5',
-      category: 'GquantLab',
-      args: args
-    });
-    palette.addItem({
-      command: 'gquant:aspect0.7',
-      category: 'GquantLab',
-      args: args
-    });
-    palette.addItem({
-      command: 'gquant:aspect1',
-      category: 'GquantLab',
-      args: args
-    });
-    palette.addItem({
-      command: 'add:outputCollector',
+      command: COMMAND_ADD_OUTPUT_COLLECTOR,
       category: 'Add New Nodes',
       args: args
     });
     palette.addItem({
-      command: 'gquant:convertCellToFile',
+      command: COMMAND_CONVERT_CELL_TO_FILE,
       category: 'GquantLab',
       args: args
     });
     palette.addItem({
-      command: commandClean,
+      command: COMMAND_CLEAN,
       category: 'GquantLab',
       args: args
     });
     palette.addItem({
-      command: commandExecute,
+      command: COMMAND_EXECUTE,
       category: 'GquantLab',
       args: args
     });
   }
   //add key board shortcuts
   app.commands.addKeyBinding({
-    command: 'gquant:reLayout',
+    command: COMMAND_RELAYOUT,
     keys: ['Alt A'],
     selector: '.jp-GQuant'
   });
   app.commands.addKeyBinding({
-    command: 'add:outputCollector',
+    command: COMMAND_ADD_OUTPUT_COLLECTOR,
     keys: ['Alt O'],
     selector: '.jp-GQuant'
   });
   //add key board shortcuts
   app.commands.addKeyBinding({
-    command: commandExecute,
+    command: COMMAND_EXECUTE,
     keys: ['Alt R'],
     selector: '.jp-Notebook'
   });
   app.commands.addKeyBinding({
-    command: commandClean,
+    command: COMMAND_CLEAN,
     keys: ['Alt C'],
     selector: '.jp-Notebook'
   });
   app.commands.addKeyBinding({
-    command: 'gquant:reLayout',
+    command: COMMAND_RELAYOUT,
     keys: ['Alt A'],
     selector: '.jp-Notebook'
   });
   app.commands.addKeyBinding({
-    command: 'add:outputCollector',
+    command: COMMAND_ADD_OUTPUT_COLLECTOR,
     keys: ['Alt O'],
     selector: '.jp-Notebook'
   });
 
   //add key board shortcuts
   app.commands.addKeyBinding({
-    command: commandExecute,
+    command: COMMAND_EXECUTE,
     keys: ['Alt R'],
     selector: '.jp-LinkedOutputView'
   });
   app.commands.addKeyBinding({
-    command: commandClean,
+    command: COMMAND_CLEAN,
     keys: ['Alt C'],
     selector: '.jp-LinkedOutputView'
   });
   app.commands.addKeyBinding({
-    command: 'gquant:reLayout',
+    command: COMMAND_RELAYOUT,
     keys: ['Alt A'],
     selector: '.jp-LinkedOutputView'
   });
   app.commands.addKeyBinding({
-    command: 'add:outputCollector',
+    command: COMMAND_ADD_OUTPUT_COLLECTOR,
     keys: ['Alt O'],
     selector: '.jp-LinkedOutputView'
   });
 
-  commands.addCommand('gquant:openAnNewNotebook', {
-    label: 'Convert TaskGraph to Notebook',
-    caption: 'Convert TaskGraph to Notebook',
-    icon: notebookIcon,
-    execute: async () => {
-      let mainView: MainView;
-      let objStr = '';
-      if (isCellVisible()) {
-        // Prompt the user about the statement to be executed
-        mainView = getMainView();
-        objStr = JSON.stringify(
-          mainView.contentHandler.privateCopy.get('value')
-        );
-      }
-      if (isGquantVisible()) {
-        mainView = app.shell.currentWidget as any;
-        objStr = JSON.stringify(
-          YAML.parse(mainView.contentHandler.context.model.toString()),
-          null,
-          2
-        );
-      }
-      const input1 = `import json\nfrom gquant.dataframe_flow import TaskGraph\nobj="""${objStr}"""\ntaskList=json.loads(obj)\ntaskGraph=TaskGraph(taskList)\ntaskGraph.draw()`;
-      const input2 = 'taskGraph.run(formated=True)';
-      return createNewNotebook(input1, input2);
-      // Execute the statement
-    },
-    isVisible: isGquantVisible
-  });
   app.contextMenu.addItem({
-    command: 'gquant:openAnNewNotebook',
+    command: COMMAND_OPEN_NEW_NOTEBOOK,
     selector: '.jp-GQuant'
   });
 }
